Modernization to make compilable with ghc 7.10
This breaks backwards-compatibility with GHC 6 and probably also
compatibility with other Haskell implementations. Oh well.

* hierarchical modules in most .lhs files
* change .ly file encoding from latin-1 to UTF-8 to make
  recent `happy' happy
* kill TABs
* Fix untyped Control.Exception.catch uses
* Provide simple Applicative/Functor implementation for Monads
* avoid binding to `rec' which is a reserved keyword for mdo language feature
* .gitignore

+++ tab/general/set.c
@@ -186,8 +186,8 @@ void set_print(struct set *s) {
   int i, j;
   for (i=0; i<size; i++)
     for (j=0; j<INT_BITS; j++)
-      if ((s->field[i] & 1 << j))
-        printf("%d ", (INT_BITS * i + j));
+      if ((s->field[i] & 1 << j)) 
+        printf("%ld ", (long)(INT_BITS * i + j));
   printf("\n");
 }
